feat(mapper): implement mapper 348 (BMC-830118-C)

+++ src/cartridge/mapper12.rs
@@ -114,8 +114,8 @@ impl Mapper for Mapper12 {
         self.inner.write_chr_1k_at(final_bank, offset, value);
     }
 
-    fn mapper_number(&self) -> u8 {
-        Self::MAPPER_NUMBER
+    fn mapper_number(&self) -> u16 {
+        u16::from(Self::MAPPER_NUMBER)
     }
 
     fn wram_size(&self) -> usize {
