+++ src/cartridge/mapper49.rs
@@ -127,8 +127,8 @@ impl Mapper for Mapper49 {
         self.mmc3.write_chr_1k_at(bank, offset, value);
     }
 
-    fn mapper_number(&self) -> u8 {
-        Self::MAPPER_NUMBER
+    fn mapper_number(&self) -> u16 {
+        u16::from(Self::MAPPER_NUMBER)
     }
 
     fn registers_snapshot(&self) -> Vec<u8> {
